Refactor timelinechart
Extract shared utilities which will be used in the observatory status
timeline as well. Mostly in the useEChartsTimeline hook
